+++ examples/proxyServer.ts
@@ -2,13 +2,7 @@
  * This is an example proxy server.
  */
 
-import {
-  bold,
-  cyan,
-  green,
-  red,
-  yellow,
-} from "https://deno.land/std@0.215.0/fmt/colors.ts";
+import { bold, cyan, green, red, yellow } from "jsr:@std/fmt@0.218/colors";
 
 import { Application, HttpError, proxy, Status } from "../mod.ts";