+++ request.ts
@@ -1,16 +1,19 @@
 // Copyright 2018-2024 the oak authors. All rights reserved. MIT license.
 
 import { Body } from "./body.ts";
+import { ServerSentEventStreamTarget } from "./deps.ts";
 import {
   accepts,
   acceptsEncodings,
   acceptsLanguages,
   type HTTPMethods,
+  ServerSentEventTarget,
+  type ServerSentEventTargetOptions,
   UserAgent,
 } from "./deps.ts";
 import type { ServerRequest, UpgradeWebSocketOptions } from "./types.ts";
 
-export interface OakRequestOptions {
+interface OakRequestOptions {
   jsonBodyReviver?: (key: string, value: unknown) => unknown;
   proxy?: boolean;
   secure?: boolean;
@@ -152,7 +155,7 @@ export class Request {
   /** An object representing the requesting user agent. If the `User-Agent`
    * header isn't defined in the request, all the properties will be undefined.
    *
-   * See [std/http/user_agent#UserAgent](https://deno.land/std@0.215.0/http/user_agent.ts?s=UserAgent)
+   * See [std/http/user_agent#UserAgent](https://deno.land/std@0.218.2/http/user_agent.ts?s=UserAgent)
    * for more information.
    */
   get userAgent(): UserAgent {
@@ -231,6 +234,22 @@ export class Request {
       return acceptsLanguages(this.#serverRequest, ...langs);
     }
     return acceptsLanguages(this.#serverRequest);
+  }
+
+  /** Take the current request and initiate server sent event connection.
+   *
+   * > ![WARNING]
+   * > This is not intended for direct use, as it will not manage the target in
+   * > the overall context or ensure that additional middleware does not attempt
+   * > to respond to the request.
+   */
+  async sendEvents(
+    options?: ServerSentEventTargetOptions,
+    init?: RequestInit,
+  ): Promise<ServerSentEventTarget> {
+    const sse = new ServerSentEventStreamTarget(options);
+    await this.#serverRequest.respond(sse.asResponse(init));
+    return sse;
   }
 
   /** Take the current request and upgrade it to a web socket, returning a web