+++ util.ts
@@ -191,15 +191,6 @@ export function readableStreamFromAsyncIterable(
  * When the pull algorithm is called on the stream, a chunk from the reader
  * will be read.  When `null` is returned from the reader, the stream will be
  * closed along with the reader (if it is also a `Deno.Closer`).
- *
- * An example converting a `Deno.FsFile` into a readable stream:
- *
- * ```ts
- * import { readableStreamFromReader } from "https://deno.land/std/io/mod.ts";
- *
- * const file = await Deno.open("./file.txt", { read: true });
- * const fileStream = readableStreamFromReader(file);
- * ```
  */
 export function readableStreamFromReader(
   reader: Deno.Reader | (Deno.Reader & Deno.Closer),
@@ -399,8 +390,13 @@ export function encodeBase64Safe(data: string | ArrayBuffer): string {
   return encodeBase64(data).replace(/\/|\+|=/g, (c) => replacements[c]);
 }
 
+export function isBun(): boolean {
+  return "Bun" in globalThis;
+}
+
 export function isNode(): boolean {
-  return "process" in globalThis && "global" in globalThis;
+  return "process" in globalThis && "global" in globalThis &&
+    !("Bun" in globalThis) && !("WebSocketPair" in globalThis);
 }
 
 export function importKey(key: Key): Promise<CryptoKey> {